code - change pair type of front+back select to string-int

+++ app/src/main/java/com/tegaoteam/application/tegao/ui/learning/cardcreate/fragment/CardCreateSetFrontFragment.kt
@@ -61,7 +61,7 @@ class CardCreateSetFrontFragment: Fragment() {
             } else {
                 _parentViewModel.submitSelectedFront(selected.map {
                     val split = it.id.split("#")
-                    split.first() to split.last()
+                    split.first() to split.last().toInt()
                 })
                 when (_parentViewModel.selectedType) {
                     LearningCardConst.Type.TYPE_FLASHCARD.id -> {
